fix: 예외 케이스 Test 분리
