+++ src/test/java/lotto/ApplicationTest.java
@@ -1,7 +1,6 @@
 package lotto;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -51,33 +50,6 @@ class ApplicationTest extends NsTest {
     void 예외_테스트() {
         assertSimpleTest(() -> {
             runException("1000j");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @DisplayName("1000원에 딱 맞아떨어지는 비용이 지불되지 않으면 예외가 발생한다.")
-    @Test
-    void notdividedByThousandWon() {
-        assertSimpleTest(() -> {
-            runException("200");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @DisplayName("1000원에 딱 맞아떨어지는 비용이 지불되지 않으면 예외가 발생한다.")
-    @Test
-    void payZero() {
-        assertSimpleTest(() -> {
-            runException("0");
-            assertThat(output()).contains(ERROR_MESSAGE);
-        });
-    }
-
-    @DisplayName("보너스 넘버는 로또 넘버와 같으면 안된다.")
-    @Test
-    void checkBonusNum() {
-        assertSimpleTest(() -> {
-            runException("8000", "1,2,3,4,5,6", "6");
             assertThat(output()).contains(ERROR_MESSAGE);
         });
     }
