+++ src/test/java/lotto/LottoTest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LottoTest {
@@ -27,5 +25,4 @@ class LottoTest {
     }
 
     // 아래에 추가 테스트 작성 가능
-
 }
